hello-next-js - Tasks API - client-side front-end layer
part 2
