--- myapps/hello-next-js/pages/bff-tasks-db/index.tsx
@@ -1,45 +1,0 @@
-import Layout from "../../pagesLayouts/Layout";
-import { GetServerSideProps, InferGetServerSidePropsType } from 'next';
-import { SWRConfig } from 'swr';
-import { TaskModel } from '@/app/models/Task/use-client/TaskModel';
-import { TaskPage } from '@/app/views/Task/from-client/taskPage';
-//import { Task } from '@/app/types/Task';
-import { DATA_FETCH_MODE } from '../../feature-flags/tasksBff';
-
-//type Props = {
-//  fallback: Record<string, Task[]>;
-//};
-
-// dev note: alternatively...
-//const BffTasksDB = ({ fallback }: Props) => {
-const BffTasksDB = ({ fallback }: InferGetServerSidePropsType<typeof getServerSideProps>) => {
-  const body = (
-    <Layout title="hello-next-js - Next.JS API - Backend For Frontend (BFF) demo">
-      <TaskPage />                 
-      <br />
-    </Layout>
-  );
-  
-  return DATA_FETCH_MODE === "getServerSideProps" 
-    ? <SWRConfig value={{ fallback }}>{body}</SWRConfig>
-    : body;
-};
-
-export const getServerSideProps: GetServerSideProps = async () => {
-  if (DATA_FETCH_MODE === "getServerSideProps") {
-    const tasks = await (new TaskModel()).getTasksDBRows();
-    return {
-      props: {
-        fallback: {
-           'Tasks-API': tasks,
-        },
-      },
-    };
-  }
-
-  return {
-    props: {}, // skip SSR, that means getTasksDBRows() is called with useEffect (CSR approach)
-  };
-};
-
-export default BffTasksDB;
