+++ myapps/hello-next-js/pages/task-crud-fullstack/index.tsx
@@ -1,0 +1,14 @@
+import Layout from "../../pagesLayouts/Layout";
+import { CLIENT_SIDE_FRONTEND_LAYOUT_TITLE } from "../../pagesLayouts/layout-title";
+import { TaskPage } from '@/app/views/Task/from-client/taskPage';
+
+const TaskList = () => {
+  return (
+    <Layout title={CLIENT_SIDE_FRONTEND_LAYOUT_TITLE}>
+      <TaskPage />                 
+      <br />
+    </Layout>
+  );  
+};
+
+export default TaskList;
