+++ myapps/hello-next-js/pages/task-crud-fullstack/swagger.tsx
