+++ myapps/hello-next-js/pagesLayouts/layout-title.ts
@@ -1,0 +1,4 @@
+const layoutTitlePartial = "Hello-Next-JS | Full stack demo |";
+
+export const CLIENT_SIDE_FRONTEND_LAYOUT_TITLE = `${layoutTitlePartial} client-side frontend ('use client') with BFF to access the backend API`;
+export const SERVER_SIDE_FRONTEND_LAYOUT_TITLE = `${layoutTitlePartial} server-side frontend ('use server') direct data fetching to backend API`;
