+++ myapps/hello-next-js/src/app/page.tsx
@@ -15,8 +15,8 @@ export default function Home() {
     mvvmSagaRedux: { title: "MVVM Pattern/Redux Saga", href:"mvvm-redux-saga" },
     exampleMySharedUI: { title: "Example/My Shared UI package", href:"example-my-shared-ui" },
     formExample: { title: "Example/Form with useActionState", href:"form-example" },
-    bffTasksPostgreDB: { title: "BFF/Tasks CRUD with PostgreSQL", href:"bff-tasks-db" },
-    bffTasksGraphQL: { title: "BFF/Tasks CRUD with PostgreSQL (Apollo GraphqQL)", href:"bff-tasks-db/graphql" },   
+    TaskListFullStackDemoMVVMClient: { title: "FullstackDemo/Tasks CRUD with remote RDB - client-side MVVM with BFF", href:"task-list-fullstack" },
+    TaskListFullStackDemoGraphQLMVVMClient: { title: "FullstackDemo/Tasks CRUD with remote RDB (Apollo GraphqQL) - client-side MVVM with BFF", href:"task-list-fullstack/graphql" },   
   }
 
   const renderedLinks: ReactElement[] = [];
